requirements.txt
tensorflow
numpy
opencv-python
easyocr
mss
Pillow
psutil
keyboard
